Moved ExceptionHandlerConfig
